+++ Tests/TestHelpers.swift
@@ -47,6 +47,14 @@ extension GitHub: TargetType {
             return "{\"login\": \"\(name)\", \"id\": 100}".data(using: String.Encoding.utf8)!
         }
     }
+
+    var validate: Bool {
+        return true
+    }
+
+    var headers: [String: String]? {
+        return nil
+    }
 }
 
 func url(_ route: TargetType) -> String {
@@ -94,6 +102,10 @@ enum HTTPBin: TargetType {
             return "{\"authenticated\": true, \"user\": \"user\"}".data(using: String.Encoding.utf8)!
         }
     }
+
+    var headers: [String: String]? {
+        return nil
+    }
 }
 
 public enum GitHubUserContent {
@@ -136,6 +148,9 @@ extension GitHubUserContent: TargetType {
         }
     }
 
+    public var headers: [String: String]? {
+        return nil
+    }
 }
 
 private let defaultDownloadDestination: DownloadDestination = { temporaryURL, response in