+++ Changelog.md
@@ -1,4 +1,10 @@
 # Next
+- **Breaking Change**: Added support to get the response (if any) from `MoyaError`.
+- **Breaking Change** Added support to get the response (if any) from `MoyaError`.
+- **Breaking Change** Updated `RxMoyaProvider.request` to return a [`Single<Request>`](https://github.com/ReactiveX/RxSwift/pull/1123).
+- **Breaking Change** Changed `Moya.Response`'s `response`to use an `HTTPURLResponse` instead of a `URLResponse`.
+- Updated the `RxSwift` version requirement to `3.3`.
+- **Breaking Change** Added `headers` to `TargetType`.
 
 - Enabled the "Allow app extension API only" flag.
 - `ReactiveSwiftMoyaProvider` and `RxMoyaProvider`'s `init`s have been deprecated.